[Modify] 자격증 요소 수정 및 삭제 시 본인확인

+++ back/src/services/CertificateService.js
@@ -1,4 +1,4 @@
-import { Certificate } from '../db'; // from을 폴더(db) 로 설정 시, 디폴트로 index.js 로부터 import함.
+import { Certificate, User } from '../db'; // from을 폴더(db) 로 설정 시, 디폴트로 index.js 로부터 import함.
 import { v4 as uuidv4 } from 'uuid';
 import { format } from 'express/lib/response';
 
@@ -38,14 +38,20 @@ class CertificateService {
     static async setCertificate({ certificate_id, toUpdate }) {
         // 우선 해당 id 의 수상내역이 db에 존재하는지 여부 확인
         let certificate = await Certificate.findById({ certificate_id });
-        const keys = Object.keys(toUpdate);
-        const values = Object.values(toUpdate);
 
         // db에서 찾지 못한 경우, 에러 메시지 반환
         if (!certificate) {
             const errorMessage = '삭제되었거나 등록되지 않은 자격증입니다.';
             throw new Error(errorMessage);
         }
+
+        if (certificate.user_id !== user_id) {
+            const errorMessage = '수정할 수 없습니다.';
+            throw new Error(errorMessage);
+        }
+
+        const keys = Object.keys(toUpdate);
+        const values = Object.values(toUpdate);
 
         // 업데이트 대상에 title이 있다면, 즉 title 값이 null 이 아니라면 업데이트 진행
         for (let i = 0; i < keys.length; i++) {
@@ -59,7 +65,13 @@ class CertificateService {
         return certificate;
     }
 
-    static async deleteCertificate({ certificate_id }) {
+    static async deleteCertificate({ certificate_id, user_id }) {
+        const certificate = await Certificate.findById({ certificate_id });
+
+        if (certificate.user_id !== user_id) {
+            const errorMessage = '삭제할 수 없습니다.';
+            throw new Error(errorMessage);
+        }
         await Certificate.deleteById({ certificate_id });
         return;
     }
